Unmanaged Memory Migliorare il rilascio della memoria inutilizzata

+++ mai/network/SoftmaxCrossEntropyLoss.cs
@@ -5,11 +5,11 @@ namespace mai.network
     public class SoftmaxCrossEntropyLoss
         : Loss
     {
-        private readonly double epsilon;
+        private readonly float epsilon;
         private bool singleOutput;
         private Matrix softmaxPrediction;
 
-        public SoftmaxCrossEntropyLoss(double epsilon = 0.000000001)
+        public SoftmaxCrossEntropyLoss(float epsilon = 0.000000001f)
         {
             this.epsilon = epsilon;
             singleOutput = false;
@@ -41,8 +41,8 @@ namespace mai.network
 
             softmaxPrediction = prediction.Softmax(epsilon, 1 - epsilon);
 
-            Matrix softmaxCrossEntropy = softmaxPrediction.Log().Hadamard(-1d * target) -
-                                         (1d - softmaxPrediction).Log().Hadamard(1d - target);
+            Matrix softmaxCrossEntropy = softmaxPrediction.Log().Hadamard(-1f * target) -
+                                         (1f - softmaxPrediction).Log().Hadamard(1f - target);
 
             return softmaxCrossEntropy.Sum() / prediction.Rows;
         }
